telux: Add APIs to perform HTTP transaction on AP for add profile functionality.
Add APIs to send response for HTTP request from AP back
to modem and also listener API to listen for incoming
HTTP requests.

This consist of http daemon for sending a http request
and send the response back to modem. For adding profile
using AP HTTP data stack, the modem sends the HTTP request
in form of indication to HTTP Daemon. Then HTTP daemon
performs HTTP post request on behalf of modem and get the
response which is forwarded to modem and finally modem
installs profile on the eUICC.

Update documentation for getProfileId() API

CRs-Fixed: 2838767

Change-Id: I5768368d9d1773976fafe64b0399031b641a08fc