telux: Allowing unlock card by puk in PIN not blocked state.
Current implementation allows unlock card by puk option only
for an app state in puk mode because of which we are not able to
perform unlock by puk when PIN is not blocked.
Add a fix to allow unlock card by puk by removing the app state
check for PUK and allowing for all app states.

CRs-Fixed: 2827007

Change-Id: I1466452e2116b47b1f735570ad027797f683b438