telux: Fix to incorrect call list reports in test app after SSREQ
Issue:
Consider a voice call is brought up after enabling audio through the
options in DialerMenu in telsdk_console_app. When SSREQ is issued
while the call is in progress, telsdk_console_app enters a state
where it stops receiving any solicited/unsolicited responses from RIL.

It is caused by the the indefinite blocking of the listener callback
(onCallInfoChange()) in telsdk_console_app, caused by unnecessary and
indefinite wait for stopAudio() callback.

Fix:
Generally a callback is expected from SDK only when the synchronous status
of the API is successful. If the synchronous status is a failure, the
application should not expect/wait for a callback.

CRs-Fixed: 2903553

Change-Id: Id62545760e21b168ecb376428121785abdf35f24