telux: Fix missing indications to data connection listener
Issue: The data connection listener is registered only after the
sub-system comes up. It is possible that certain indications that
are triggered during the sub-system init are lost in this case.

Fix: Register listener as soon as the manager is created instead
of waiting until the sub-system is available

CRs-Fixed: 2966699

Change-Id: I533f46884b88178f20f43ab977d2a242516289bd