telux: Adding device configurabilty in ecall_app
Issue: Currently test teams are not able to configure
the audio device for ecall app, since the app was using the
string for device that are defined in public header,
and since the devices are configurable using tel.conf
it was leading to invalid mappings of stream to device.

Fix: Changed the input format for device in config file
from string to integer value so that the device is also
configurable.

CRs-Fixed: 3034897

Change-Id: Ifcf45ec6389da61bd46e5e7a12cd09c2358035ff