telux: fixes asnbuf memory mishanlding in wsmp and creates codec lib
wsmp library is allocating an asn buffer for the purpose of encoding
messages. This change removes that unnecessary allocation, creates
a way to continue encoding with un-allocated abuf and abuf buffer
memory and makes qits allocate the wsmp abuf and its buffer memory
on sae application initialization

V2x codec that is used in the reference apps now is built as a
public library to be used in the system.

Change-Id: Id54fe0f6f5c52d5034b2c6df1aef13ee02977cb0