telux: Fixing segmentation fault of sensor_test_app
- Whenever a sensor MLC feature is enabled, on cleanup
of sensor feature manager, sensor test app crashes with
segfault. This was because, while iterating the set of
features, the iterator pointing to the feature was removed
causing next iterator to compare with an invalid pointer.

Fix - Updating the iterator by retrieving it via the erase
function instead of incrementing directly.

CRs-Fixed: 3359295, 3369875

Change-Id: Ib8dd5850f7f8a949031ee3cb0f799c7e0da75b48