audio: fix for race condition during SSR in console_app
A race condition can happen if an operation is performed
on a stream and SSR occurs. When SSR occurs cleanup function
tries to release all the resources in one thread which
could lead to seg fault in console app if an audio operation
is in progress on another thread.

To prevent this, adding mutex to synchronise the cleanup
thread and the audio operation thread.

Change-Id: Id27576d6ae9848d838d5a7b39b59b974bb21e42f