data: VLAN Hypervisor Virtual Interface Revised
With this change vmtap0 virtio is used by both tele-vm and fota-vm
to create virtual interface to communicate to pvm.
pvm uses virtual interface vmtap0 to communicate with tele-vm and
virtual interface vmtap1 to communicate with fota vm.
In addition, when telsdk_console_app is running in pvm, available
virtual interfaces displayed were changed to the  following
vmtap-televm to refer to vmtap0
vmtap-fotavm to refer to vmtap1

Change-Id: I56ec45b7c9a16d5d693d3ecb474f18aabd1a456e