common: Handle error states for cin
When user inputs are piped to any console app, the cin stream
of the console enters failed state after all the inputs are
written on the write end of the pipe and post this only EOF
is being written. Since, piping redirects the inputs to cin
once EOF is reached cin stream always reads EOF and eofbit is
set for cin. When we again try to read, the failbit is also set
indicating read operation failed to complete successfully.

Getline extraction stops if the end of file is reached in stream
or delimitation character is found. When cin's eofbit is set and
we try to read using getline the extraction stop with 0 bytes read.
This results in an infinte loop.

To avoid this, added cin stream-state check to exit gracefully
when cin receives eof or enters a bad state.

CRs-Fixed: 3575047
Change-Id: Id9e08d728bd519e1eeb46c9e80b086862ea8312f