loc: Exclude ENGINE_NMEA from default report mask
Issue- Clients saw a failure when starting
Detailed engine reports. This failure occurs when the client
does not provide the optional GnssReportTypeMask parameter OR
when the client provides a value where the mask has the bits
for ENGINE_NMEA and NMEA both set.
For example the failure occurs if the client passes
DEFAULT_GNSS_REPORT, as the value for the reportMask.

Fix- Removed the ENGINE_NMEA bit from the DEFAULT_GNSS_REPORT
value used by the reportMask parameter in the
startDetailedReports and startDetailedEngineReports
as a default parameter.

CRs-Fixed: 3604543

Change-Id: I4bdd1ee68275c91d09d22b1eb36e6b15d4cc4614