loc: Fix for Default Gnss Report Mask.
The initial fix provided for GNSS Report mask
was with the #define Macro. However if clients use
the macro with the telux::loc:: namespace, the Macro
replacement during pre-processing converts into a
compilation error because of the pre-pended symbols.

Fix- Change the macro to a constant and define below
the Enum definition.

CRs-Fixed: 3604543

Change-Id: I63254ad32931ec30f342e46a40cc316f86558d4b