loc: Updating location test app for external data injection.
Issue- During auto injection, location data parameters are
expected to be populated on a different thread.
Since the main thread is also active,
input for location data would be considered as
invalid on the main thread.

Fix- Populate the location data parameters on the main thread
and pass the updated info to the detached thread.

CRs-Fixed: 3949244

Change-Id: I4cc8bb3276387f4a18e00d7d8aa27ae30bc194fb