power: Notify clients about the state update.
Issue - SDK library doesn't relay the updated state
information to clients for the registered machine.
So, in the case of SSR during suspend, the system resumes
but this updated information is not sent to clients.

Fix - Add an API to inform the clients about the
current state of the registered machine whenever the
power management subsystem is available.

CRs-Fixed: 4153157

Change-Id: I33f0d65845adf989c38fdf9815a4d0602ceb6596